requirements.txt
alembic==1.13.2
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytest==8.2.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.31
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
